+++ Sources/EspressoMartiniCLI/main.swift
@@ -1,0 +1,1 @@
+EspressoMartiniCLI.main()